Followup for make TokuDB compile with GCC-8
Missed printfs from: 21246066b296bd9c6e44272a8f93cb7dc73ea43c

+++ storage/tokudb/PerconaFT/ft/ft-ops.cc
@@ -826,7 +826,7 @@ int toku_ftnode_fetch_callback(CACHEFILE UU(cachefile),
                 __FILE__,
                 __LINE__,
                 toku_cachefile_fname_in_env(cachefile),
-                blocknum.b);
+                (longlong)blocknum.b);
         } else {
             fprintf(
                 stderr,
@@ -836,7 +836,7 @@ int toku_ftnode_fetch_callback(CACHEFILE UU(cachefile),
                 __FILE__,
                 __LINE__,
                 toku_cachefile_fname_in_env(cachefile),
-                blocknum.b,
+                (longlong)blocknum.b,
                 r);
         }
         // make absolutely sure we crash before doing anything else.
